refactor: déplacement de quelques types dans shared [skip ci]
