+++ shared/constants/indicateurs.ts
@@ -1,0 +1,10 @@
+export const typesEffectifNominatif = [
+  "apprenant",
+  "apprenti",
+  "inscritSansContrat",
+  "rupturant",
+  "abandon",
+  "inconnu",
+] as const;
+
+export type TypeEffectifNominatif = (typeof typesEffectifNominatif)[number];
